fix(deploy): strip v2-only options before calling v1 path
